pyseane/requirements.txt
beautifulsoup4>=4.12.2
Django>=4.2.7
django_environ>=0.11.2
Requests>=2.31.0
certifi>=2021.10.8
urllib3[socks]<3,>=1.26
selenium>=4.15.2
webdriver_manager>=4.0.1